Clearer installation instructions
I've found at least 2 people plus myself, struggling a bit to find the instructions. So I think it would a good idea to move them to the README.md in a super concise way. We can keep all the extras in the FAQ.

This is a proposal. Feel free to either keep the changes or modify them to your disposal.
Cheers.